[native] Fix broken StyleProps type for react-native-reanimated libdef
Summary:
This issue was first introduced in D5556, and led to a `$FlowFixMe` being introduced in D13066.

It's not possible to spread a non-object. I remove the spreads here because the only relevant part of the union is the object form (`{ [name: string]: any, ... }`), which is already included in the `StyleProps` type anyways.

This diff addresses part of [ENG-9039](https://linear.app/comm/issue/ENG-9039/investigate-flow-errors).

Test Plan: Flow

Reviewers: inka, tomek

Subscribers: