[lib] distinguish between Inbound and Outbound `DMOperationSpecification`
Summary:
[ENG-8423](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages).

In `useProcessDMOperation` we need to know if this operation is generated on this device (where code is executed) or received from the other peer.

This is also an enhancement to better flag when to use `MessageSourceMetadata` (used for sending confirmation).

Test Plan:
Flow

Note: This stack is tested end-to-end according to the process described [here](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages#comment-b409cae7).

Reviewers: tomek, inka