[lib] update `processDMOpsActionType` to handle `outboundP2PMessages`
Summary:
[ENG-8423](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages).

We need to process DM Operations and save `OutboundP2PMessages` in DB in **the same SQL transaction** to handle cases when the app crashes after applying changes to our store and before addressing messages to other peers - this is inconsistency from which we're unable to heal.

Depends on D13146

Test Plan:
Tested in next diff.

Note: This stack is tested end-to-end according to the process described [here](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages#comment-b409cae7).

Reviewers: tomek, inka

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13148