[keyserver] Specify MariaDB port in backups code
Summary:
In D12336 we added support to `DBConfig` for a port, but we forgot to update the backups code to specify this port when calling `mysqldump`.

This addresses at least part of [ENG-9077](https://linear.app/comm/issue/ENG-9077/investigate-broken-mariadb-backup-cron-script).

Test Plan: Ran this on my keyserver instance running in a Docker container on my personal server and confirmed that it works

Reviewers: will

Reviewed By: will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13134