[native] Fetch only a subset of messages from thick threads
Summary:
For the thick threads we don't want to initially fetch all the messages. We're going to fetch them in batches while the user scrolls.

https://linear.app/comm/issue/ENG-8702/modify-nativedatasqlite-data-handlerjs-to-fetch-only
https://linear.app/comm/issue/ENG-8703/modify-webshared-workerutilsstorejs-to-fetch-only

Depends on D13133

Test Plan: On web, disable pruning completely, create a thick thread with 100 messages. Close and open the app and check if only 20 messages were fetched.
With the pruning enabled, modify the query to fetch 5 messages and check if thick threads don't contain more.

Reviewers: kamil, inka