[lib] move generating P2P messages to `useProcessDMOperation`
Summary:
[ENG-8423](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages).

We need to process DM Operations and save `OutboundP2PMessages` in DB in **the same SQL transaction** to handle cases when the app crashes after applying changes to our store and before addressing messages to other peers - this is inconsistency from which we're unable to heal.

The existing place of generating messages is removed later in the stack.

Depends on D13148

Test Plan:
Check that messages are persisted along with changes to store.

Note: This stack is tested end-to-end according to the process described [here](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages#comment-b409cae7).

Reviewers: tomek, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13149