[lib] deprecate `sendDMOperation` and implement method to create sending promise
Summary:
[ENG-8423](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages).

We still need promise to be resolved after sending messages, but generating messages was moved to `processDMOps`.

Returning promise to later in that stack put it to `dispatchActionPromise`.

`useProcessAndSendDMOperation` looks no obsolete, but will be updated later in the stack.

Depends on D13150

Test Plan:
Call `getDMOpsSendingPromise` and check if it works.

Note: This stack is tested end-to-end according to the process described [here](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages#comment-b409cae7).

Reviewers: tomek, inka