[lib] add `outboundP2PMessageIDs` to `LocalMessageInfo` to track progress
Summary:
[ENG-8423](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages).

We decided to use the existing store for local messages for two reasons, to avoid adding a new store, and because in components, `LocalMessageInfo` is easily accessible (I think it saves us at least a couple of days of work).

Additionally `LocalMessageInfo` seems like a reasonable place to put this data.

All fields in `LocalMessageInfo` are optional, which means this does not require a migration.

Depends on D13152

Test Plan:
Make sure `outboundP2PMessageIDs` are persisted & read on app start.

Note: This stack is tested end-to-end according to the process described [here](https://linear.app/comm/issue/ENG-8423/support-e2ee-text-messages#comment-b409cae7).

Reviewers: tomek, inka