[native] Expose messages fetching function through JSI
Summary:
Expose the function through JSI so that it can be used on native clients.

https://linear.app/comm/issue/ENG-8701/create-c-functions-that-fetch-a-subset-of-messages

Depends on D13140

Test Plan: Tested this later in the stack by creating a thick thread on native, with 100 messages, closing and reopening the app, and scrolling up - verified that all the messages were eventually shown.

Reviewers: kamil, inka

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13141