[lib] Fetch thick thread messages
Summary:
Use the same mechanism as we use for the thin threads.

The change in message reducer is necessary because persisting the information about `startReached` status to SQLIte causes a bug where after refreshing the app we don't try to fetch messages ever again. An alternative solution would be to filter the status out on the DB layer, but I think this one is a lot cleaner.

https://linear.app/comm/issue/ENG-8706/modify-message-lists-to-call-the-new-function

Depends on D13147

Test Plan:
Created a thick thread with 100 messages. Closed and reopened the app, scrolled up, and made sure all 100 messages are displayed.
There are two issues though:
1. We don't stop after fetching the first batch. Instead, we're fetching all the batches one by one immediately.
2. The robotext about creating a thread isn't fetched when there are more than `defaultNumberPerThread` messages.

Reviewers: kamil, inka