Return single curve25519 string from NotificationsInboundKeysProvider
Summary: This differential returns only curve25519 key string from native code and creates JSON in C++ with folly instead of creating JSON in native code. The rieason for this change is that we saw issues with the content of the JSON returned from Java - Android app was unable to create session as inbound since sender identity key was corrupted. This change fixed the issue.

Test Plan:
1. Before this diff creating notifs session with Android device as inboud would fail each (> 90%) of the times.
2. After this diff it always works.

Reviewers: kamil

Subscribers: