Enrich CommMMKV api with operations that manipulate string set
Summary: This differential enriches CommMMKV api with new methods that add, remove and get size of set of strings stored in MMKV under some key. The purpose here is to keep list of unread thick threads in MMKV and updated it and use it from notifs code to handle the bacge icon update.

Test Plan: Tested in next differential

Reviewers: bartek

Subscribers: