[lib] implement adding thread members for DMs
Summary:
[ENG-8568](https://linear.app/comm/issue/ENG-8568/logic-for-adding-a-user-user-joining-a-thick-thread)

This hook which wraps two needed operations together, `add_viewer_to_thread_members` and `add_members`.

Doing this in parallel, not at once is bad, created a task to fix this: [ENG-9103](https://linear.app/comm/issue/ENG-9103/update-processandsenddmoperation-to-support-multiple-dm-ops-at-once), as this is complex change.

Depends on D13182, D13193

Test Plan: Tested later in the stack

Reviewers: tomek, inka

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13197