[web] Fix modal styles
Summary:
issue: [[https://linear.app/comm/issue/ENG-8679/scroll-in-pinned-messages-on-web-seems-to-be-broken | ENG-8679]]
The issue was that the pinned messages modal div was not scrollable, but fixing this revealed another issue: contents of Modal were wrapped by `modalContentContainer` which allowed its content to extend it.

Test Plan:
Checked ALL places where Modal component is used. Checked that after these changes they all work correctly. Here are modals that got changed:
{F2568094}

{F2568093}

{F2568092}

{F2568090}

{F2568086}

Here is before and after for the search component (using prod and dev):
{F2568087}

Reviewers: tomek, ashoat

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13191