[lib] Update thread and keyserver ID regex
Summary:
This regex is used on web to recognize thread id in the url. When it isn't matched, we navigate to a default thread.

https://linear.app/comm/issue/ENG-9099/opening-a-link-to-a-thick-thread-resets-the-navigation-on-web

Test Plan: Open a web page on a thin thread that isn't the first, refresh the page and check if the same thread is still open.

Reviewers: kamil, inka