[keyserver] only wrap migration in transaction if specified in migrationType
Summary:
If the migration type is `wrap_in_transaction_and_block_requests`, we wrap the migration in a transaction. Only if it was wrapped in transaction do we run rollback if the query failed.

Depends on D13209

Test Plan: I console logged and made sure that transaction and rollback queries were only being run if the migration in `migrations` specified wrapping in a transaction.

Reviewers: varun, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13212