[lib] Introduce thread timestamps on the JS side
Summary:
Create a new property with the timestamps. All the new thick threads will have it. We don't really need to have a migration introducing them, because all the thick threads we have were created by code on dev environments.

https://linear.app/comm/issue/ENG-8751/add-a-timestamps-property-to-a-thick-thread-info

Test Plan: On native and web, create a new thick thread, check if it contains the `timestamps`. Restart the apps and check if the timestamps are still there.

Reviewers: kamil, inka