[lib][web][native] Update createRealThreadFromPendingThread to handle thick thread creation
Summary:
issues:
[[https://linear.app/comm/issue/ENG-8442/update-createrealthreadfrompendingthread-to-handle-thick-thread | ENG-8442]], [[https://linear.app/comm/issue/ENG-9122/address-todos-in-d13160 | ENG-9122]]

Test Plan:
Tested by setting `usingOlmViaTunnelbrokerForDMs` to `true`. Not setting this flag yet, because it has other effects.
Tested that a thick thread gets created on both web and native.

Tested that if the flag is set to `false`, thin threads are created correctly

There is a problem with navigation - on both platfrms when the thick thread is created the user is not navigate to it. Instead they stay in the pending thread. I will address that in [[https://linear.app/comm/issue/ENG-9147/fix-navigation-in-thick-thread-creation | ENG-9147]]

Reviewers: tomek, kamil

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13244