[keyserver] pin the rust version we use in our keyserver docker image
Summary: resolves https://linear.app/comm/issue/ENG-9162/docker-keyserver-ci-is-failing-due-to-voprf-crate

currently we install the latest stable version of rust in our keyserver docker image. unfortunately, opaque-ke 2.0 has a dependency that's not compatible with the latest version of rust anymore. we'll need them to cut a new release before we upgrade our rust version. in the meantime, we'll pin the version of rust used in our services, since we know this version is compatible with opaque-ke 2.0.

Test Plan: https://github.com/vdhanan/comm/actions/runs/10725583006/job/29743742109 successful CI run on my fork of comm repo

Reviewers: ashoat, will

Subscribers: