[native] Bump react-native-screens to 3.29.0
Summary: This is the latest version that I was able to get to work with our setup. More details [here](https://linear.app/comm/issue/ENG-8812/update-android-target-api).

Test Plan: CI and tested running the Android app on simulator. Before releasing this I'll test a release build on physical iOS and Android devices and play around with it a bit

Reviewers: varun, will

Reviewed By: will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13250