[native] Bump Android target SDK version to 34
Summary: This addresses [ENG-8812](https://linear.app/comm/issue/ENG-8812/update-android-target-api). We need to bump this in order to continue submitting new versions of the Comm Android app to the Play Store.

Depends on D13250

Test Plan: CI and tested running the Android app on simulator. Before releasing this I'll test a release build on a physical Android device and play around with it a bit

Reviewers: varun, will, marcin

Subscribers: