[lib] mark messages as sent when there is no `messageID` in Outbound messages table
Summary: Handle edge-case when the message was sent, but it wasn't updated in the message store (e.g. app was killed), but was already confirmed by the other peer - and when hitting retry there is no message with given ID in the outbound messages table.

Test Plan: Test case described in summary.

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13255