[lib] mark messages as sent when there is no `messageID` in Outbound messages table
Summary: Handle edge-case when message was sent, but it wasn't updated in the message store (e.g. app was killed), but was already confirmed by the other peer - and when hitting retry there is no message with given ID in outbound messages table.

Test Plan: Test case described in summary.

Reviewers: tomek

Subscribers: