[lib][native] Review usingOlmViaTunnelbrokerForDMs and pendingThreadType
Summary:
issue: [[https://linear.app/comm/issue/ENG-9146/review-usingolmviatunnelbrokerfordms-and-pendingthreadtype | ENG-9146]]
`pendingThreadType` should actually return thick or thin types depending on where we call it from.
We want to keep the `usingOlmViaTunnelbrokerForDMs` flag, so that for now thick threads are not created.

Test Plan: Tested that with the flag set to `true` thick threads can be created. Checked that with the flag set to `false`, thin threads are created instead.

Reviewers: tomek, kamil

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13258