[web] Avoid keyserver call when deleting media for DMs
Summary:
On web, we pre-upload media as soon as it is selected, before actually sending the message. When user decides not to send the image (clicks the cross button in the corner),
the uploaded media gets deleted. This results in keyserver call: `this.props.deleteUpload()`. For DMs, we want to avoid such call.

Needed to update some code to pass the whole `threadInfo` of active chat, instead of only `threadID`.

Depends on D13281

Test Plan:
- Verified that behavior for keyserver media hasn't been changed and the endpoint is still called
- TBD for DMs

Reviewers: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13282