[lib] Update existingThreadInfoFinder to handle thick threads
Summary:
issue: [[https://linear.app/comm/issue/ENG-8433/rethink-existingthreadinfofinder | ENG-8433]]
If a chat under GENESIS exists for the given list of users, we want to return it. Otherwise we want to check for DM chat for the given set of users.

Test Plan:
Tested that is a chat under GENSIS exists, it is returned. Tested that otherwise, if a DM chat exists, it is returned. Tested that if neither exist, no chat is shown, and it is possible to create a chat (DM or under GENESIS, depening on `usingOlmViaTunnelbrokerForDMs` flag).

Reviewers:
tomek, kamil
Subscribers: