[terraform] bump prod identity to 0.35
Summary:
we probably can't deploy this until monday at this point, but preparing the diff now so that we're not blocked on getting it approved later.

confirmed that @ashoat tagged and pushed 0.35 to dockerhub after the apparent issues pushing 0.35-staging to dockerhub yesterday.

won't land until we're ready to deploy

Test Plan: terraform plan

Reviewers: will, ashoat, bartek

Reviewed By: will

Subscribers: tomek, ashoat

Differential Revision: https://phab.comm.dev/D13252