[identity] avoid returning errors if one-time keys are missing
Summary: in our olm fork, one-time keys are now optional. however, in the identity service, we still have a couple places where we treat them as required, and fail certain requests if we're unable to fetch them. we should update these parts of the code to complete RPCs successfully regardless of OTK retrieval success or failure.

Test Plan:
- Manually deleted OTKs from my test account on staging, then called `GetKeyserverKeys` and `GetOutboundKeysForUser` and got successful responses from both (without any OTKs in the response)
- Modified `get_one_time_key` to return an error and confirmed that both RPCs were still successful

Reviewers: bartek, kamil, will

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13278