[identity] avoid returning errors if one-time keys are missing
Summary: in our olm fork, one-time keys are now optional. however, in the identity service, we still have a couple places where we treat them as required, and fail certain requests if we're unable to fetch them. we should update these parts of the code to complete RPCs successfully regardless of OTK retrieval success or failure.

Test Plan: will fill in shortly

Reviewers:

Subscribers: