[lib] Fix useProcessAndSendDMOperation
Summary:
This addresses [[https://linear.app/comm/issue/ENG-9187/local-message-stays-in-new-dms | ENG-9187]], as well as fixes a more general issue with `useProcessAndSendDMOperation`:
If the callback from `useProcessAndSendDMOperation` was called twice in a continuous block of code, the second call would not see the effects of the first call. This is because underneath, `dispatchActionPromise` is called that changes the state. So for the second call to see the effects of the first call, a render has to happen between the calls.

Test Plan:
Tested that when creatiing a new thick thread, the message appears to be in progress and then sent / failed.

Reviewers:
tomek, kamil
Subscribers: