[lib] Allow sending operations to some devices
Summary:
Introduce a possibility to choose to which devices a message should be sent. We need this so that primary device can share its state with a just added device.

https://linear.app/comm/issue/ENG-9240/allow-sending-p2p-messages-to-one-device-in

Test Plan: Log in on 3 different devices as the same user. Create a thread creation operation on one device. Send it to one device and verify that it was received only on it.

Reviewers: kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13292