[keyserver] modify community fetcher to return channel tag
Summary: rather than a boolean indicating whether the channel tag exists or not, we should return the actual tag. this way, we can check if the channel ID is in the list of channels led by the joining user

Test Plan: will test with next diff. will not land until next diff is accepted

Reviewers: ashoat

Subscribers: