[lib] Set failedQueryCacheTimeout to 0 in UserIdentityCache
Summary: This addresses [ENG-9255](https://linear.app/comm/issue/ENG-9255/reduce-failedquerycachetimeout-in-useridentitycache), and makes it so if a user identity fetch times out, the next time it's requested, we don't return `null` from the cache, but instead retry the fetch.

Test Plan: I set `queryTimeout` to 1 second, which made it time out. I confirmed in the logs that after the timeout, a second request was dispatched

Reviewers: varun, will

Reviewed By: varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13299