[lib] delete `CallSingleKeyserverEndpointResultInfoInterface`
Summary:
[ENG-8424](https://linear.app/comm/issue/ENG-8424/update-inputstatecontainer-to-support-multimedia-messages-being-sent).

Later in the stack, we need to return `SendMessagePayload` (which has ` +interface: CallSingleKeyserverEndpointResultInfoInterface`) as a result of DM Ops. which is encrypting and queueing messages on Tunnelbroker, not keyserver call.

Alternatives:
1. Use `tunnelbroker` instead of `none` - I think this is even worse for readability because the type is associated with Keyserver.
2. Remove `interface` from `SendMessagePayload` - this seems the best and I think it's safe because I wasn't able to find any usage of `interface`.

I am not familiar with why we need this at all, so curious about the reviewers' perspective.

Depends on D13319

Test Plan: Flow

Reviewers: tomek, inka, ashoat

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13321