[lib] Share the state with a new device
Summary:
The primary device shares sends all the thick threads to a new device.

https://linear.app/comm/issue/ENG-9241/introduce-a-handler-that-sends-operations-to-new-devices

Test Plan: Create a new account. Create some thick threads. Log in on a new device and check if the threads appear. Close the primary, log in on a new device, open the primary - the threads should appear.

Reviewers: kamil, inka

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13307