[lib] implement hook to send composable DMOperation
Summary:
[ENG-8424](https://linear.app/comm/issue/ENG-8424/update-inputstatecontainer-to-support-multimedia-messages-being-sent).

Improve documenting code and addresses https://phab.comm.dev/D13290#inline-76570, we have a separate and clear hook for composable messages. This will replace some of the code, which is removed later in the stack.

Depends on D13326

Test Plan: Tested later in the stack (both content and notifs)

Reviewers: tomek, inka

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13327