[native] Launch SIWE backup secret creation
Summary:
While the backup itself isn't ready, we're ready to generate the secret for SIWE backup. It would be good to get this info from users as part of the new registration flow.

Depends on D13309

Test Plan: Make sure SIWE backup screen shows up in registration flow when using an ETH wallet

Reviewers: marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13310