[native] Fix Comm swoosh render in RegistrationTerms on iOS release builds
Summary:
For some reason, `center` works on iOS debug builds, but not release builds. This diff makes it look good on release builds.

`center` is probably closer to what we want than `contain`, but they have the same behavior because the swoosh is larger than the screen width and as such doesn't need to be centered.

Depends on D13311

Test Plan:
| before | after |
| | |

Reviewers: varun, will

Subscribers: