[native] Fix handleCurrentUserFID querying for current user's FID
Summary: This addresses [ENG-9230](https://linear.app/comm/issue/ENG-9230/cant-see-option-to-manage-farcaster-tags-on-web). The cause was a simple error from D12569. Flow failed to catch it because it wasn't sure that `currentUserID` wasn't the string `"identities"`.

Test Plan: I deployed a release build to my iOS device and confirmed that the Profile -> Farcaster account screen now showed a linked Farcaster account.

Reviewers: varun, will

Reviewed By: varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13313