[web] Fix broken pending threads
Summary:
In D13282 we modified the logic so that instead of an active thread ID we use active thread info that is taken from the thread store. Unfortunately, this change broke the pending threads which aren't present in the store. What we should do instead is to let the component provide the thread store for us.

https://linear.app/comm/issue/ENG-9268/pending-threads-are-broken-on-master-on-the-web

Test Plan: Check if it is possible to create a new thread by sending a text message in a pending thread.

Reviewers: kamil, bartek, ashoat