[lib][native][web] send encrypted dms on web and native when editing entries on thick threads
Summary:
This uses e2ee dm ops with the save entry hook instead of the keyserver when sending entries in a thick thread

Depends on D13343

Test Plan: two accounts each with a web app and iOS simulator. Ensured on all four platforms/devices that editing entries showed between the four platforms when sending entries in a thick thread.

Reviewers: marcin, ashoat, tomek, kamil

Reviewed By: tomek, kamil

Differential Revision: https://phab.comm.dev/D13344