[web] Remove admins tab when there are. no admins
Summary:
In thick threads there are no admins, so it doesn't make sense to show an empty tab. Instead of hiding the tab just for thick threads, we can hide the tab every time there are no admins.

https://linear.app/comm/issue/ENG-9258/remove-admins-tab-in-thick-threads

Test Plan: Open a members menu in a thread with some admins and check if both tabs appear. Open it in a thick thread and check if only the members tab is present.

Reviewers: kamil, bartek, ashoat