[lib] have edit entry message infos use e2ee op timestamp instead of creationTime
Summary:
Addresses out of order edit and create robotexts. I introduced a bug by setting the message info timestamp to the `creationTime` instead of the time the edit op was created.

This results in out of order robotext for creating and editing entries. The fix is to use the op creation timestamp instead of `creationTime` which is when the entry was created.

Test Plan:
Tested on iOS simulator. entries are back in order.

Reviewers:
ashoat,
tomek,
marcin

Subscribers: