[lib] refactor methods from `PeerToPeerContext` to utils
Summary: [ENG-9106](https://linear.app/comm/issue/ENG-9106/in-createolmsessionswithpeer-support-creating-multiple-sessions-with).

The code in Context is messy and complicated, moving this to methods to make it clean. See usages in next diff - I am going to link source code from Githb to make it easier for review.

Depends on D13373

Test Plan: Tested in TBD

Reviewers: bartek!, tomek!, marcin