[web] make it possible to select farcaster avatar on web
Summary:
if the user has an fid, we try to get their farcaster avatar URL. if we find a URL, we'll display the option to set the farcaster avatar as the user's comm avatar. if the user selects the farcaster avatar, their avatar is updated in the app and on the keyserver.

Depends on D13379

Test Plan: can successfully set farcaster avatar on web and it subsequently appears on other clients. can reset the avatar, too.

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13380