[native][web] make farcaster avatars work as thread avatars
Summary: in order to set the farcaster avatar as thread avatar in 1 to 1 chats, we need to get the other user's FID from state.

not sure why we have so much code duplication here, but i'd like to defer addressing it.

Depends on D13380

Test Plan: thread avatars now look right when the other person in chat has set a farcaster avatar

Reviewers: ashoat, tomek

Subscribers: