[lib] Introduce findUserIdentities to ProcessDMOperationUtilities
Summary:
This goes through `UserIdentityCache`, so we are guaranteed that multiple fetches for the same user ID will result in at most a single identity call.

Depends on D13389

Test Plan: For now, just Flow. I'll test this at a later point in the stack; for now, putting up diffs following the team's recent pattern to get feedback

Reviewers: tomek, kamil, will

Differential Revision: https://phab.comm.dev/D13390