[keyserver] new fetcher to get all communities
Summary:
we need the community id, name, and farcaster channel id for all communities so we can display them to users in a modal.

part of https://linear.app/comm/issue/ENG-9216/new-keyserver-endpoint-to-get-a-list-of-communities

Test Plan: tested the SQL query directly on my local mariadb instance. will test further in subsequent diffs that call the keyserver endpoint that returns this data

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D13396