[lib] Avoid unnecessary sets creation
Summary: We don't need to create sets for the queues - we can as effectively check the object keys. The benefit of the new approach is that we don't perform a linear amount of work when new entities appear.

Test Plan:
1. Created an operation with a reaction that was received before an operation creating the message
2. Created an operation with an entry edit that was received before an operation creating the entry
3. Created an operation with thread subscription change that was received before the user became a member

In each of the cases verified that the result was correct and that the operation was removed from the queue.

Reviewers: kamil, will

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13400