[lib] If a viewer isn't a member, don't assign them the role
Summary:
When we create a thick thread, we should check if the viewer is a member and set the role only when they are. This case can happen when we receive a create sidebar operation.

https://linear.app/comm/issue/ENG-9318/sidebar-joining-issues

Test Plan: Create a thick sidebar without the viewer and check if the join button is visible.

Reviewers: kamil, will