[keyserver][lib] Introduce update_relationships2 keyserver endpoint
Summary:
We'll update the client code to use this new keyserver endpoint in the next diff. I did this refactor as part of my work on [ENG-9331](https://linear.app/comm/issue/ENG-9331/add-a-flag-to-update-relationships-to-disable-keyserver-threadrobotext).

Depends on D13439

Test Plan: I tested this new endpoint as part of my testing of [ENG-9269](https://linear.app/comm/issue/ENG-9269/relationship-buttons-result-in-a-thin-thread-being-created)

Reviewers: tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13440