[lib] Include targetUserID in DMUpdateRelationshipOperation
Summary:
This was an oversight from D13392. We need to process the same DM op on both the sender and recipient, so we should include a `targetUserID` in the op so it's clear who the target is. Otherwise, the sender sees "you sent a friend request to you".

Depends on D13432

Test Plan: Confirm the sender sees the proper robotext after this change

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D13433