[lib] don't process persisted inbound messages without `viewerID`
Summary:
[ENG-9381](https://linear.app/comm/issue/ENG-9381/seeing-ignored-dmoperation-because-logged-out).

I had some `InboundP2PMessage`'s in the store, and I discovered that they're causing `ignored DMOperation because logged out`.
This is a race condition because this component can run before getting `loggedInUserInfo` in `getInitialReduxState` and `useProcessDMOperation` was called without `viewerID`.

This fixes the issue.

This is also safe for messages not from DB but received from Tunnelbroker, we're going to establish socket connection only when `viewerID` is defined anyway.

Test Plan: Seeing the messages processed.

Reviewers: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13445