[keyserver][lib] Allow skipping createRobotextInThinThread
Summary:
This parameter only needs to be specified for `FARCASTER_MUTUAL` and `FRIEND`. By making it optional, we can address [this feedback](https://phab.comm.dev/D13443?id=44476#inline-77071) on D13443.

Submitting it as a separate diff as I'm not sure the complexity is worth it, or whether there are other alternatives that reviewers might prefer.

Depends on D13443

Test Plan: Flow, and tested the relationship adding flow end-to-end again

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D13451