[lib] fix `useThreadSearchIndex` to avoid ignoring private thread
Summary:
[ENG-9378](https://linear.app/comm/issue/ENG-9378/cant-find-genesis-private-thread-when-searching-ones-username#comment-57c320a9).

This code was ignoring private threads and as a result, causing that private thread not to show up.
On prod, it works only for old users where by default thread name was the user's name, after changing the name in settings it is not working anymore. Right now, when creating an account `name` is optional and not set to own username - that is way this is not working for dev where we have new users.

Test Plan: Searching for my username works (even if I changed name of my private thread).

Reviewers: tomek, ashoat