[lib][native] fix farcaster outline icon and update CommIcons.ttf
Summary:
icon was off center and as a result looked a little strange on web. it also wasn't an outline like the other icons in the list. lastly, i forgot to include the updated CommIcons.ttf file, which we need to make the icon available on native

this diff re-centers the icon, makes it an outline, and updates CommIcons.ttf

Test Plan:
icon looks correct now. CommIcons.ttf update is tested in next diff which actually uses it to display farcaster icon on native (android)

also tested that some of the existing CommIcon icons (e.g. "cloud-filled") still look right on native

{F2810494}

Reviewers: will

Reviewed By: will

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13475