[lib][native] fix farcaster outline icon and update CommIcons.ttf
Summary: icon was off center and as a result looked a little strange on web. also, i forgot to include the updated CommIcons.ttf file, which we need to make the icon available on native

Test Plan: icon looks correct now. CommIcons.ttf update is tested in next diff which actually uses it to display farcaster icon on native (android)

Reviewers: will

Subscribers: