[native] add farcaster icon on native
Summary:
missed this in my previous diff stack. we need to add the farcaster icon to `icons` to make it available in the action sheet.

i also resized the ethereum-outline icon because it looked too small

Depends on D13475

Test Plan:
see attached photos. farcaster icon now appears in action sheet and the ethereum icon looks right

{F2810499}

{F2810498}

(simply replaced the icon name to test the ethereum icon size, not actually using that icon for the farcaster avatar option in action sheet)

Reviewers: will

Reviewed By: will

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13476